\name{cd}
\alias{cd}
\title{Shorthand for setwd, remembering previous directory
A setwd replacement, allowing \code{cd()} to be like \code{cd -} on unix (return to last dir)}
\usage{
  cd(dir)
}
\arguments{
  \item{dir}{path to new directory}
}
\description{
  Shorthand for setwd, remembering previous directory A
  setwd replacement, allowing \code{cd()} to be like
  \code{cd -} on unix (return to last dir)
}

