% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox_folder.R
\name{dropbox_folder}
\alias{dropbox_folder}
\title{Find the Dropbox folder on a Windows machine}
\source{
This solution comes from \url{http://stackoverflow.com/questions/35985167/determining-the-dropbox-path-in-r}
}
\usage{
dropbox_folder()
}
\value{
The path to the Dropbox folder, a character string
}
\description{
Sourcing functions or reading data files can be a problem when they are stored in a Dropbox
folder you access on different machines.  This function uses the Dropbox file,
\code{info.json}, always found at either \code{\%APPDATA\%\\Dropbox\\info.json} or
\code{\%LOCALAPPDATA\%\\Dropbox\\info.json}
}
\examples{
\dontrun{
.Dropbox <- dropbox_folder()
my_file <- paste(.Dropbox, "R/my_file.csv", sep="")
mydata <- read.csv(my_file)
}
}
