% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls.objects.R
\name{ls.objects}
\alias{ls.objects}
\title{ls with added information on objects}
\source{
Peter Pikal, from R-help 8-2-2016
}
\usage{
ls.objects(pos = 1, pattern, order.by)
}
\arguments{
\item{pos}{position in the search list}

\item{pattern}{an optional regular expression. Only names matching pattern are returned.}

\item{order.by}{one of \code{"Type"}, \code{"Size"}, \code{"Rows"}, \code{"Columns"}}
}
\value{
a data frame, with columns Type, Size, Rows, Columns. Object names are the row names.
}
\description{
\code{ls} with added information on object Type, Size, Rows, and Columns
}
\examples{
ls.objects()
ls.objects(pattern="*lm")
ls.objects(order.by="Type")
}
