% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_checks.R
\name{release_check_win}
\alias{release_check_win}
\title{Check a package on win-builder (R-devel)}
\usage{
release_check_win()
}
\value{
invisible NULL
}
\description{
Check a package on win-builder (R-devel)
}
