% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_checks.R
\name{release_site}
\alias{release_site}
\title{Rebuild a package's README and pkgdown site}
\usage{
release_site()
}
\value{
invisible NULL
}
\description{
Rebuild a package's README and pkgdown site
}
\seealso{
Other release checks:
\code{\link[=release_build]{release_build()}},
\code{\link[=release_check]{release_check()}},
\code{\link[=release_check_win]{release_check_win()}},
\code{\link{release_checks}},
\code{\link[=release_cran_comments]{release_cran_comments()}},
\code{\link[=release_document]{release_document()}},
\code{\link[=release_preflight]{release_preflight()}},
\code{\link[=release_revdep]{release_revdep()}},
\code{\link[=release_run_all]{release_run_all()}},
\code{\link[=release_spelling]{release_spelling()}},
\code{\link[=release_spelling_add]{release_spelling_add()}},
\code{\link[=release_urls]{release_urls()}}
}
\concept{release checks}
