% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myutil.R
\name{short.path}
\alias{short.path}
\title{Return a shortened version of a path}
\usage{
short.path(dir, len = 2)
}
\arguments{
\item{dir}{path to new directory}

\item{len}{number of final path components to return}
}
\description{
Return a shortened version of a path
}

