% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_github.R
\name{update_github}
\alias{update_github}
\title{update_github}
\usage{
update_github()
}
\value{
none
}
\description{
Update installed Github packages
}
\details{
Find the local packages that were installed from Github and update them
}
\author{
From: https://stackoverflow.com/questions/32538052/update-all-packages-from-github
}
