% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_data_doc.R
\name{use_data_doc}
\alias{use_data_doc}
\title{Generate Outline Documentation for a Data Set in Roxygen Format}
\usage{
use_data_doc(object, filename = NULL, name = NULL, labels = NULL)
}
\arguments{
\item{object}{an R object to be documented as a data set}

\item{filename}{usually, a \code{connection} or a character string giving the name of the file
to which the documentation shell should be written. The default corresponds to a file
whose name is \code{name} followed by \code{".R"}. Can also be \code{NA} (see details below).}

\item{name}{a character string specifying the name of the object. Defaults to the name of \code{object}.}

\item{labels}{a character vector of variable labels or strings describing each column in the data set.}
}
\value{
If \code{filename} is \code{NA}, a list-style representation of the documentation shell.
Otherwise, the name of the file written to is returned invisibly.
}
\description{
Generates a shell of documentation for a data set or other object in roxygen format.
This function was created by editing \code{\link[utils]{promptData}} to replace
the old style \code{.Rd} formatting with code suitable for processing with \code{\link[devtools]{document}}.
}
\details{
Unless \code{filename} is \code{NA}, a documentation shell for \code{object} is written in roxygen format to the file specified
by \code{filename}, and a message about this is given.

If \code{filename} is \code{NA}, a list-style representation of the documentation shell is created
and returned. Writing the shell to a file amounts to \code{cat(unlist(x), file = filename, sep = "\n")},
where \code{x} is the list-style representation.

Currently, only data frames are handled explicitly by the code.
}
\examples{
\dontrun{
use_data_doc(iris)
unlink("iris.R")

# using variable labels
labels <- c("Sepal length (mm)", "Sepal width (mm)", "Petal length (mm)", "Petal width (mm)", "Iris species" )
# console output
zz <- use_data_doc(iris, filename=NA, labels=labels)
cat(unlist(zz), sep="\n")
}

}
\seealso{
\code{\link[utils]{promptData}}
}
\author{
Michael Friendly
}
