% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man.R
\name{man}
\alias{man}
\title{Find and display documentation for a package
Opens the CRAN documentation for a package or builds it from the local R library}
\usage{
man(package, method = c("system", "web"))
}
\arguments{
\item{package}{name of a package, with or without quotes}

\item{method}{either \code{"web"} or \code{"system"}.  \code{"web"} uses the manual from CRAN.
\code{"system"}  builds the documentation from your local R version.}
}
\description{
Find and display documentation for a package
Opens the CRAN documentation for a package or builds it from the local R library
}
