\name{myutil-package}
\alias{myutil-package}
\alias{myutil}
\docType{package}
\title{
My R Utility Functions for Rgui

}
\description{
Designed to be loaded by .Rprofile, this provides some utilities
I like to have available in Rgui.

}
\details{
\tabular{ll}{
Package: \tab myutil\cr
Type: \tab Package\cr
Version: \tab 1.5\cr
Date: \tab 2013-08-02\cr
License: \tab GPL-2\cr
}

}
\author{
Michael Friendly

Maintainer: Michael Friendly <friendly@yorku.ca>
}
%\references{
%~~ Literature or other references for background information ~~
%}
\keyword{ package }
\seealso{
\code{\link{eps}}, \code{\link{man}}

\code{\link{setwd}}, \code{\link{cd}}
}
\examples{
# None
}
