% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_checks.R
\name{release_revdep}
\alias{release_revdep}
\title{Check a package's reverse dependencies}
\usage{
release_revdep(num_workers = 4)
}
\arguments{
\item{num_workers}{number of parallel workers to use}
}
\value{
invisible NULL
}
\description{
One-time setup if \code{revdepcheck} isn't installed:
\code{remotes::install_github("r-lib/revdepcheck")}
}
