% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb2col.R
\name{rgb2col}
\alias{rgb2col}
\title{Find Named Color Closest to Given Hex RGB Colors
This function calculates the distance (in CIE Lab color space) between
the input colors (in RGB hex notation) and all named \code{colors()}
It returns the closest named color within a given threshold distance.}
\source{
From J. Fox, R-help, 5-30-2013
}
\usage{
rgb2col(cols, near = 2.3)
}
\arguments{
\item{cols}{A vector of colors, in RGB hex notation, e.g., \code{"\#00AE30"}}

\item{near}{Threshold for nearest color from \code{colors()}.  The default corresponds to 1 JND in Lab space.}
}
\value{
A vector of color names, or the hex value if no color is close enough
}
\description{
Find Named Color Closest to Given Hex RGB Colors
This function calculates the distance (in CIE Lab color space) between
the input colors (in RGB hex notation) and all named \code{colors()}
It returns the closest named color within a given threshold distance.
}
\examples{
cols <- c("#010101", "#EEEEEE", "#AA0000", "#00AA00", "#0000AA", "#AAAA00", "#AA00AA", "#00AAAA")
(nms <- rgb2col(cols))
pie(rep(1, 2*length(cols)), labels=c(rbind(cols, nms)), col=c(rbind(cols, nms)))
}

