% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myutil.R
\name{setwd}
\alias{setwd}
\title{Set Working Directory and Display in Window Title}
\usage{
setwd(dir)
}
\arguments{
\item{dir}{path to new directory}
}
\description{
Set Working Directory and Display in Window Title
}

